mveg-rag/requirements.txt
flask
pypdf
sentence-transformers
faiss-cpu
numpy